% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdj_seq.R
\name{add_nt_and_aa_clonotypes}
\alias{add_nt_and_aa_clonotypes}
\title{Clonotype by CDR3 amino acid or nucleotide seq}
\usage{
add_nt_and_aa_clonotypes(
  clonotypes_and_contigs_list = list(clonotypes, filtered_contig_annotations)
)
}
\arguments{
\item{clonotypes_and_contigs_list}{[Required] A list of length == 2. The first element is a tibble data frame of clonotypes (derived from the cellranger clonotypes.csv file). The second element is a tibble data frame of contig annotations (derived from the cellranger filtered_contig_annotations.csv file).}
}
\value{
A list of length 2: list[[1]] are the clonotypes; list[[2]] are the contigs.
}
\description{
Use the cellranger-determined clonotypes (based on full nt seq of VDJ region) and the CDR3 amino acid sequences to add additional clonotype classifications to each original clonotype. Two or more cellranger (raw) clonotypes might have the same exact CDR3 amino acid sequence -- this function would give these different raw clonotypes the same "cdr3aa" clonotype id. The same process is done for CDR3 nt sequences.
}
\examples{
\dontrun{
clonotypes <- read_csv("outs/clonotypes.csv", col_types = cols(.default = "c"))
filtered_contig_annotations <- read_csv("outs/filtered_contig_annotations.csv", col_types = cols(.default = "c"))
out <- tk_add_nt_and_aa_clonotypes(list(clonotypes, filtered_contig_annotations))
}

}
