% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annot.R
\name{check_gene_db}
\alias{check_gene_db}
\title{Check gene names against NCBI database and fix nomenclature}
\usage{
check_gene_db(gene_vector, gene_db)
}
\arguments{
\item{gene_vector}{Required. Character vector of input gene names. These can be slightly incorrect in their capitalization or be gene aliases.}

\item{gene_db}{Required. Tibble data frame of the GENE database downloaded from NCBI. See below for details.

How to download gene_db:
url <- "https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Mammalia/Homo_sapiens.gene_info.gz"
url <- "https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Mammalia/Mus_musculus.gene_info.gz"
download.file(url, basename(url), method = "wget")
gene_db <- read_tsv(gzfile(basename(url)))}
}
\value{
None
}
\description{
Check gene names against NCBI database and fix nomenclature
}
\examples{
\dontrun{
genes_clean <- check_gene_db(genes_dirty, gene_db)
}

}
