% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_mem}
\alias{clean_mem}
\title{Release memory}
\usage{
clean_mem(n = 10)
}
\arguments{
\item{n}{Integer, Number of times to run garbage collection. Default is 10.}
}
\value{
None
}
\description{
Run after deleting large objects to initiate garbage collection multiple times.
[Reference](https://stackoverflow.com/a/1467334/2367748)
}
\examples{
\dontrun{
clean_mem()
}

}
