% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkoutdir}
\alias{mkoutdir}
\title{Make an output directory using scratch location}
\usage{
mkoutdir(out_dir, use_scratch = FALSE)
}
\arguments{
\item{out_dir}{String, path to output directory}

\item{use_scratch}{Logical. Should the output dir be created on scratch, with a symlink to the scratch location placed at out_dir?}
}
\value{
None
}
\description{
Make an output directory using scratch location
}
\examples{
\dontrun{
mkoutdir()
}

}
