% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_calls.R
\name{robust_system}
\alias{robust_system}
\title{System calls and stderr stdout reporting}
\usage{
robust_system(command)
}
\arguments{
\item{command}{[Required] String that is passed directly to the base::system() command.}
}
\value{
A list of length 3. list[[1]] is the exit code for the command. list[[2]] is the "stdout" text. list[[3]] is the "stderr" text.



This function was mostly copied from:
[https://stackoverflow.com/a/17090601/2367748](https://stackoverflow.com/a/17090601/2367748)
}
\description{
Robustly pass commands to system() and capture stderr and stdout data
}
\examples{
robust_list <- robust_system("date; cal")

}
\seealso{
[robust_system_check()]
}
