% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_calls.R
\name{robust_system_check}
\alias{robust_system_check}
\title{Print the data captured by robust_system() and stop if nonzero exit code}
\usage{
robust_system_check(
  robust_list = NULL,
  name = deparse(substitute(robust_list))
)
}
\arguments{
\item{robust_list}{[Required] List object that is returned by the [robust_system()] function.}

\item{name}{[Optional. Default: name of `robust_list` object] String that provides a name for the system call captured in the `robust_list`.}
}
\value{
This function prints the contents of the `robust_list` in a nice way, directly to stderr and stdout streams. It also checks the exit code and will `stop()` if not zero.
}
\description{
Print the data captured by robust_system() and stop if nonzero exit code
}
\examples{
robust_system_check(robust_list)

}
\seealso{
[robust_system()]
}
