% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_rnaseq.R
\name{tk_PercentAbove}
\alias{tk_PercentAbove}
\title{tk_PercentAbove
https://github.com/satijalab/seurat/blob/4e868fcde49dc0a3df47f94f5fb54a421bfdf7bc/R/utilities.R#L2127-L2136
This was copied exactly from Seurat. The function is not exported from Seurat, so I included it here.
Calculate the percentage of a vector above some threshold}
\usage{
tk_PercentAbove(x, threshold)
}
\arguments{
\item{x}{Vector of values}

\item{threshold}{Threshold to use when calculating percentage}
}
\value{
Returns the percentage of `x` values above the given threshold
}
\description{
tk_PercentAbove
https://github.com/satijalab/seurat/blob/4e868fcde49dc0a3df47f94f5fb54a421bfdf7bc/R/utilities.R#L2127-L2136
This was copied exactly from Seurat. The function is not exported from Seurat, so I included it here.
Calculate the percentage of a vector above some threshold
}
